cpu_state: don't load_cpu state from const values
Currently when we generate code for loading the CPU state when entering
a translation block, we initialize each register/variable to a const
(static) value, the value that the register has at the time of the
translation.

This means that when we re-use a cached block, each time we enter it we
reload the register values it had when it was first translated. Oops.

Change load_cpu_state() method to generate loads from memories instead
of setting register/variables to constant values.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>