jit: only support hotswap/pic on x86_64
Cranelift JIT doesn't support PIC or Hotswapping (i.e. redefining
functions) on non-x86_64 architectures. Unfortunately this means we need
to re-instantiate the JIT context from scratch if we want to re-compile
a function starting from the same instruction. This happens when we want
to turn from normal execution to single step or when a breakpoint is
added and the translated entry block containing the breakpoint address
gets invalidated.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>