memory/region: add mmapped memory region support for macos
On linux, we create an memfd and then mmap it to RAM.

On macos, an mmap must be mapped as an anonymous mmapping with a special
JIT mmap flag, and we must also call pthread_jit_write_protect_np() to
turn off the write protect in the JIT area.

We should also turn it on after finishing writing.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>