gdb: add breakpoints support
Add breakpoint support, which utilises the translated block invalidation
implemented in previous commits.

Without block invalidation, the breakpoint could be in the middle of a
translated block and we could never really stop when we reach that
instruction.

Now, the block gets removed and re-translated, and we check if we need
to stop at the appropriate instruction of the breakpoint.

On breakpoint removal, the block gets invalidated again, to force
re-translation of the potentially non-partitioned block that was
previously split.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>