jit: add translate_operand_extended() method
For instructions with extended register semantics: (ADD, ADDS, SUB,
SUBS, CMP) we need to know the width we are shift-extending a register
operand. This information is not available in translate_operand()
method, so add a new one. The shift-extended operation should only be
present in those instructions, so panic!() in translate_operand if an
extended operation is encountered.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>